+++ aardwolf/_version.py
@@ -1,5 +1,5 @@
 
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 __banner__ = \
 """
 # aardwolf %s
