+++ aardwolf/examples/aardpcapscan.py
@@ -7,8 +7,8 @@ import logging
 import json
 
 from aardwolf import logger
-from aardwolf.examples.scancommons.targetgens import *
-from aardwolf.examples.scancommons.internal import *
+from aardwolf.examples.scancommons.targetgens import ListTargetGen, FileTargetGen
+from aardwolf.examples.scancommons.internal import EnumResult, EnumResultStatus
 from aardwolf.utils.univeraljson import UniversalEncoder
 
 from aardwolf.network.selector import NetworkSelector
