+++ aardwolf/examples/aardpscreenshot.py
@@ -6,9 +6,8 @@ import logging
 from aardwolf import logger
 from aardwolf.commons.url import RDPConnectionURL
 from aardwolf.commons.iosettings import RDPIOSettings
-from aardwolf.examples.scancommons.targetgens import *
-from aardwolf.examples.scancommons.internal import *
-from aardwolf.examples.scancommons.utils import *
+from aardwolf.examples.scancommons.targetgens import ListTargetGen, FileTargetGen
+from aardwolf.examples.scancommons.internal import EnumResult, EnumResultStatus, EnumProgress
 from aardwolf.commons.queuedata.constants import MOUSEBUTTON, VIDEO_FORMAT
 from tqdm import tqdm
