+++ aardwolf/protocol/pdu/capabilities/__init__.py
@@ -3,34 +3,34 @@ import enum
 from os import stat
 
 from aardwolf import logger
-from aardwolf.protocol.pdu.capabilities.bitmamachachehostsupp import *
-from aardwolf.protocol.pdu.capabilities.bitmap import *
-from aardwolf.protocol.pdu.capabilities.bitmapcache import *
-from aardwolf.protocol.pdu.capabilities.bitmapcache2 import *
-from aardwolf.protocol.pdu.capabilities.brush import *
-from aardwolf.protocol.pdu.capabilities.bitmapcodecs import *
-from aardwolf.protocol.pdu.capabilities.colortablecache import *
-from aardwolf.protocol.pdu.capabilities.compdesk import *
-from aardwolf.protocol.pdu.capabilities.control import *
-from aardwolf.protocol.pdu.capabilities.font import *
-from aardwolf.protocol.pdu.capabilities.frameack import *
-from aardwolf.protocol.pdu.capabilities.gdiplus import *
-from aardwolf.protocol.pdu.capabilities.general import *
-from aardwolf.protocol.pdu.capabilities.glyph import *
-from aardwolf.protocol.pdu.capabilities.input import *
-from aardwolf.protocol.pdu.capabilities.largepointer import *
-from aardwolf.protocol.pdu.capabilities.multifragmentupdate import *
-from aardwolf.protocol.pdu.capabilities.ninegrid import *
-from aardwolf.protocol.pdu.capabilities.offscreen import *
-from aardwolf.protocol.pdu.capabilities.order import *
-from aardwolf.protocol.pdu.capabilities.pointer import *
-from aardwolf.protocol.pdu.capabilities.remoteprograms import *
-from aardwolf.protocol.pdu.capabilities.share import *
-from aardwolf.protocol.pdu.capabilities.sound import *
-from aardwolf.protocol.pdu.capabilities.surface import *
-from aardwolf.protocol.pdu.capabilities.virtualchannel import *
-from aardwolf.protocol.pdu.capabilities.windowactivation import *
-from aardwolf.protocol.pdu.capabilities.windowlist import *
+from aardwolf.protocol.pdu.capabilities.bitmamachachehostsupp import TS_BITMAPCACHE_HOSTSUPPORT_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.bitmap import TS_BITMAP_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.bitmapcache import TS_BITMAPCACHE_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.bitmapcache2 import TS_BITMAPCACHE_CAPABILITYSET_REV2
+from aardwolf.protocol.pdu.capabilities.brush import TS_BRUSH_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.bitmapcodecs import TS_BITMAPCODECS_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.colortablecache import TS_COLORTABLE_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.compdesk import TS_COMPDESK_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.control import TS_CONTROL_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.font import TS_FONT_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.frameack import TS_FRAME_ACKNOWLEDGE_CAPABILITYSET
+#from aardwolf.protocol.pdu.capabilities.gdiplus import *
+from aardwolf.protocol.pdu.capabilities.general import TS_GENERAL_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.glyph import TS_GLYPHCACHE_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.input import TS_INPUT_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.largepointer import TS_LARGE_POINTER_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.multifragmentupdate import TS_MULTIFRAGMENTUPDATE_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.ninegrid import TS_DRAW_NINEGRID_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.offscreen import TS_OFFSCREEN_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.order import TS_ORDER_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.pointer import TS_POINTER_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.remoteprograms import TS_REMOTE_PROGRAMS_CAP_SET
+from aardwolf.protocol.pdu.capabilities.share import TS_SHARE_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.sound import TS_SOUND_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.surface import TS_SURFCMDS_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.virtualchannel import TS_VIRTUALCHANNEL_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.windowactivation import TS_WINDOWACTIVATION_CAPABILITYSET
+from aardwolf.protocol.pdu.capabilities.windowlist import TS_WINDOW_LIST_CAP_SET
 
 
 class CAPSTYPE(enum.Enum):
